// apps/server/src/generated/prisma/models/CrmPersonaProfile.ts
export type CrmPersonaProfileModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmPersonaProfilePayload>
// ...
export type $CrmPersonaProfilePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CrmPersonaProfile"
  objects: {
    organization: Prisma.$OrganizationPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    organizationId: string
    name: string
    description: string | null
    titleKeywordsText: string | null
    customerTypeKeywordsText: string | null
    painPoints: string | null
    focusText: string | null
    avoidText: string | null
    status: string
    isDefault: boolean
    createdById: string
    createdByName: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["crmPersonaProfile"]>
  composites: {}
}